requirements.txt
kafka_python==2.0.2
psycopg2-binary==2.9.9
apache-airflow==2.9.2 
pyspark==3.4.1
unidecode==1.3.7